+++ CHANGELOG.md
@@ -4,7 +4,7 @@ https://github.com/jcorporation/myMPD/
 
 ***
 
-## myMPD v12.0.2 (not yet released)
+## myMPD v12.0.2 (2023-09-22)
 
 This is a small bug fix release.